aeval/sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aeval"
version = "0.1.0"
description = "AI Evaluation Pipeline — write evals like pytest, run them against any Ollama model"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.0",
    "rich>=13.0",
    "scipy>=1.12",
    "numpy>=1.26",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.3",
]

[project.scripts]
aeval = "aeval.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/aeval"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
